Remove full results logic from mojo_python_tests_runner.py
Since we're not running the Mojo Python tests on the Chromium
waterfall anymore, we probably don't really need the code to
support uploading the results to the flakiness dashboard.

If we decide we do want this in the future, we can either add
this back or switch to the 'typ' framework that Chromium and
Blink are switching to for running python-based tests.

R=viettrungluu@chromium.org, jamesr@chromium.org
BUG=crbug.com/402172

Review URL: https://codereview.chromium.org/652953007