Changes ordering of connecting to view_manager
New flow is to connect to WindowManager, which is responsible for
connecting to ViewManager. WindowManager gets the Embed function that
is no longer part of the ViewManager (at least the one without an a
view_id). ViewManager expects initial connection to provide internal
api for viewmanager<->windowmanager communication.

I still need to do more cleanup, but this is the initial stab at this.

Also, I nuked the demo_launcher and wm_flow_init. They are replaced
with a shell script mojo/tools/mojo_demo.sh .

BUG=none
TEST=none
R=ben@chromium.org

Review URL: https://codereview.chromium.org/671993002