Roll tools for chromium->mojo updates
This adds a set of utilities useful for updating mojo's snapshot of
various directories from chromium. The compare_deps.py script compares
the DEPS file in the mojo repo with one from the chromium repo to spot
any dependencies that might need updating. The update_from_chromium.py
script pulls in various directories from a chromium checkout and then
applies a few patches where necessary.  The intended sequence to update
from chromium is:

1.) Update a chromium checkout and make sure the rev is good-ish
2.) Run mojo/tools/roll_tools/compare_deps.py in a mojo checkout and
  fix any deps issues it identifies, commit
3.) Run mojo/tools/roll_tools/update_from_chromium.py, build, run tests
  and fix up any issues that arise
4.) Squash down to a single commit containing the chromium committish
  and land.

The process is inherently manual for now since we have to patch up a
few things to work.  When we fix up our dependencies on things like net,
we can get out of the business of patching and automate this process
further.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/674723002