Store network_service cache in memory
The network cache is not safe for usage by multiple processes at the same
time like when sky tests are running many mojo_shell instances in parallel.
To work around this we store the cache in memory. In the future we should
consider file system partitioning for mojo services.

R=davemoore@chromium.org

Review URL: https://codereview.chromium.org/679403002