Folds WindowManager2 into WindowManager
In order for us to provide security in the WM connections to the WM
are routed through the VM. This allows the VM to provide the WM with the
id space of the client. Clients are also allowed to connect directly
to the wm, but everything but Embed is ignored in this case.

BUG=none
TEST=none
R=ben@chromium.org

Review URL: https://codereview.chromium.org/694823003