Add an example of a content handler implemented in a content handler.
The application recursive_content_handler is a constant handler that
will launch itself on each call to start application. It can be used to
test content handler written in another content handler.

For example by adding the following command line to the shell:
--content-handlers=image/gif,mojo://recursive_content_handler,image/jpeg,{{URL_TO_A_GIF_FILE}}

when trying to run a jpeg file the shell will try to run the gif file
as a content handler, which will run the recursive content handler. The
recursive content handler will install a new instance of itself a the
application, and the new instance will then serve the jpeg content.

R=aa@chromium.org

Review URL: https://codereview.chromium.org/690293004