Move ChannelInfo from being an internal detail of embedder.cc to mojo/system.
I want to reuse ChannelInfo for something else. Previously, it was
internal to the implementation of the embedder API (i.e., only defined
in embedder.cc, and forward-declared in embedder.h). Now, it's
forward-declared in embedder/channel_info_forward.h (since others wanted
to forward-declare it), declared in system/channel_info.h, and defined
in system/channel_info.cc.

This is a bit convoluted, but we don't want to expose the internals of
ChannelInfo to users of the embedder API. However, we still want to
allow them to safely use custom deleters, etc. without an additional
(trivial) wrapper, so we do't want to just use void*.

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/699773002