Avoid copying PRESUBMIT.py files with the sdk
Running the presubmit checks we run in mojo when rolling the code into
chromium doesn't make much sense since the code isn't modified on the
chromium side. In practice, this is problematic because the clang-format
rules change surprisingly frequently so code that is clang-formatted in
the mojo side will not necessarily format the same way in chromium.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/702933003