Changes the DEPS file to pull down the Dart VM source.
This pulls in the most recent dev-channel release of the
Dart source. It includes the Dart Editor, but it is planned
that this will be removed in the near future, which will
reduce checkout size.

The Dart source is under src/dart/dart. The disadvantage is
that this looks a bit silly, but it means that we don't have
to create/maintain yagr (Yet Another Git Repo).

From here, I'll work on translating the Dart gyp files over
to GN.

BUG=
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/670263008