Make //mojo/edk/embedder/* pass gn check and add it to the presubmit check.
Split //mojo/edk/embedder:embedder into :embedder and :platform. The
latter doesn't depend on anything in //mojo (possibly it should be moved
to a directory of its own). The files in //mojo/edk/system depend on
:platform, but mostly not :embedder (except for
embedder/configuration.h, a recent addition that should probably be
moved to a separate target).

Still to do:
* Split the //mojo/edk/system target.
* Add //mojo/edk/system/system_impl_export.h to some target (I'd like to
  do so without allowing :platform to include files in
  :system/:embedder/etc.).
* Clean up the various export macros and MOJO_USE_SYSTEM_IMPL (AFAICT, a
  lot of this mess is to "support" building apps in the component
  build).

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/704663003