Low-level Dart bindings for Mojo functions.
These bindings are written as Dart native extensions. When we have a real embedder going two things will happen. First, the native code will be compiled into the embedding. Second, AFAIU, mojo_dart_init.cc and mojo_init.dart won't be needed anymore as those calls will instead be made from the native embedder code.

In this CL, handles are wrapped by the RawMojoHandle class. On top of it will sit a class called MojoHandle that will extend the Dart API's Stream class, and be more user friendly.

In RawMojoHandle, waitMany takes a list of integer handles. This simplifies the implementation of the "handle watcher" isolate, which will come in the CL after this one. In Dart code, the only place waitMany should be called from is the handle watcher isolate, so hopefully taking a list of ints instead of a list of RawMojoHandles isn't too controversial, here.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/709173002