Fixes bug where root_ was getting set too late
When a connection goes away we need to delete the root immediately,
otherwise we may assume the root_ is still around and if asked for we
won't get it.
Previously we the sequence was OnWillDestroyViewManagerServiceImpl
and eventually ProcessViewDeleted . Problem is immediately after
OnWillDestroyViewManagerServiceImpl asking for the Root returns
NULL. This is why the DCHECK was hitting in NotifyDrawnStateChanged(),
because we happened to get a NotifyDrawnStateChanged before
ProcessViewDeleted.

BUG=432658
TEST=ViewManagerServerAppTest.OnParentOfEmbedDisconnects
R=msw@chromium.org

Review URL: https://codereview.chromium.org/720953002