Refactoring of view manager server to allow more direct testing
The app test approach is great for testing at the mojom level, but I'm
going to be adding functionality that is hard to provide good test
coverage for at the mojom level. This patch restructures things so that
it's possible to test the view manager more directly. That is, without
really using bindings.

The test here is trivial. This is more about the refactoring. Follow
on patches will provide more tests.

R=msw@chromium.org

Review URL: https://codereview.chromium.org/725693002