Introduce command buffer control thread
This CL introduce a control thread for the GPU command buffer. We listen on the
command buffer pipe on a control thread so that we can respond to messages with
low latency. Currently, we don't have any special low-latency processing on the
control thread, but the plan is to implement sync points on the control thread.
Instead, we just forward all the messages to the driver thread by re-wrapping
the arguments in base::Callbacks.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/739493002