Two minor tweaks for ease of animations in wm
FrameController: destroy the root |view_| rather than app_view_ than
root. This way we can animate the view and subviews when closing.

WM: add the View for the window to its parent before showing. This
makes showing animations much easier. It's possible we may want to
have an animation on add, but for now this simplifies things (and
matches the aura side).

BUG=434429
TEST=none
R=erg@chromium.org

Review URL: https://codereview.chromium.org/752663002