Add mojo_system_impl target for win64.
Pepper/Mojo support in Chrome requires an implementation of mojo_system_impl
that can be used in Windows 64-bit binaries. This adds another target for
Windows 64-bit builds that can be used.

This is used as part of a larger change to add Mojo support to Pepper plugins:
https://codereview.chromium.org/701273002/

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/750563006